style: reorganize
